% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_endpoints.R
\name{alerts}
\alias{alerts}
\title{Fetches a list of alerts}
\usage{
alerts(instance)
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}.}
}
\description{
Fetches a list of alerts for events such as excessive file usage
}
