% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_package}
\alias{get_package}
\title{Get package information by name}
\usage{
get_package(package_name, repo_id = 1)
}
\arguments{
\item{package_name}{The name of the package.}

\item{repo_id}{The repository ID.}
}
\description{
Fetches package information from an RSPM repository for a package by name.
}
