% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_repo_packages}
\alias{get_repo_packages}
\title{Get all packages for a repository}
\usage{
get_repo_packages(repo_id)
}
\arguments{
\item{repo_id}{The repository ID.}
}
\description{
Get all packages for a repository
}
