% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_repo_sysreqs}
\alias{get_repo_sysreqs}
\title{Get all system requirements for a repository}
\usage{
get_repo_sysreqs(repo_id, distribution = "ubuntu")
}
\description{
Get all system requirements for a repository
}
