% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{instance}
\alias{instance}
\title{Connect to an RStudio Package Manager instance}
\usage{
instance(url, silent = FALSE)
}
\arguments{
\item{url}{string The URL of the Package Manager instance}

\item{silent}{boolean Silences the connection status message}
}
\description{
Connect to an RStudio Package Manager instance
}
