% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_endpoints.R
\name{metrics_packages}
\alias{metrics_packages}
\title{Fetches usage data from the metrics API}
\usage{
metrics_packages(instance, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}.}

\item{params}{A list of optional query parametes}
}
\description{
This API takes two possible parameters resulting in a URL of the form:
\verb{/__api__/metrics/packages?_sourceType=r&_days=7}
}
