% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos}
\alias{repos}
\title{Fetches lists of repositories}
\usage{
repos(instance, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}}

\item{params}{A list of optional parameters applied to the API query}
}
\description{
Fetches lists of repositories
}
