% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_first_transaction}
\alias{repos_first_transaction}
\title{Fetches the date of the earliest calendar transaction for a repo}
\usage{
repos_first_transaction(instance, id)
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}.}

\item{id}{An RSPM repository ID}
}
\description{
Fetches the date of the earliest calendar transaction for a repo
}
