% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_packages}
\alias{repos_packages}
\title{Fetches lists of packages for a specific repository with optional filters}
\usage{
repos_packages(instance, id, name = NULL, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}}

\item{id}{An RSPM repository ID}

\item{name}{A package name}

\item{params}{A list of optional parameters applied to the API query}
}
\description{
Fetches a single package for a specific repository with optional parameters
}
