% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_packages_sysreqs}
\alias{repos_packages_sysreqs}
\title{Fetches lists of system requirements for a specific repository and package}
\usage{
repos_packages_sysreqs(instance, id, name, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}}

\item{id}{An RSPM repository ID}

\item{name}{A package name}

\item{params}{A list of optional parameters applied to the API query}
}
\description{
Fetches lists of system requirements for a specific repository and package
}
