% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repos.R
\name{repos_sysreqs}
\alias{repos_sysreqs}
\title{Fetches list of system requirements for a specific repository}
\usage{
repos_sysreqs(instance, id, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}}

\item{id}{An RSPM repository ID}

\item{params}{A list of optional parameters applied to the API query}
}
\description{
Fetches list of system requirements for a specific repository
}
