% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{return_endpoint}
\alias{return_endpoint}
\title{The generic endpoint access function}
\usage{
return_endpoint(instance, endpoint, params = list())
}
\arguments{
\item{instance}{An RSPM instance object created with \code{instance()}}

\item{endpoint}{The API endpoint to query}

\item{params}{Optional list of query parameters}
}
\description{
builds appropriate query strings and performs requests against the supplied
RSPM instance
}
