% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources}
\alias{sources}
\title{Fetches lists of sources available in RSPM}
\usage{
sources(instance)
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}}
}
\description{
Fetches lists of sources available in RSPM
}
