% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources_packages}
\alias{sources_packages}
\title{Fetches lists of packages for a specific source with optional filters}
\usage{
sources_packages(instance, id, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}}

\item{id}{An RSPM source ID}

\item{params}{A list of optional parameters applied to the API query}
}
\description{
Fetches lists of packages for a specific source with optional filters
}
