% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources_packages_releases}
\alias{sources_packages_releases}
\title{Fetches lists of PyPI package releases for a specific package name and source with optional filters}
\usage{
sources_packages_releases(instance, id, name, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}}

\item{id}{An RSPM source ID}

\item{name}{A package name}

\item{params}{A list of optional parameters applied to the API query}
}
\description{
Fetches lists of PyPI package releases for a specific package name and source with optional filters
}
