% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{sources_transactions}
\alias{sources_transactions}
\title{Fetches lists of transactions for a given source}
\usage{
sources_transactions(instance, id, params = list())
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}.}

\item{id}{An RSPM source ID}

\item{params}{A list of optional parameters applied to the API query}
}
\description{
Fetches lists of transactions for a given source
}
