% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_endpoints.R
\name{status}
\alias{status}
\title{Fetches the status of the RStudio Package Manager server}
\usage{
status(instance)
}
\arguments{
\item{instance}{An rspm_instance object created with \code{instance()}.}
}
\description{
Fetches the status of the RStudio Package Manager server
}
