% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_package_sysreqs}
\alias{get_package_sysreqs}
\title{Get system requirements for a single package}
\usage{
get_package_sysreqs(package_name, repo_id = 1, distribution = "ubuntu")
}
\description{
Get system requirements for a single package
}
